+++ CHANGELOG.md
@@ -21,3 +21,4 @@
 - **Desync involving lost-soul charge attack**
 - **_Double autoaim range_ setting doubling range of BFG tracers**
 - **FOV effects disabling interpolation of weapon sprites**
+- **Crash when loading WADs with empty lumps between `C_#` markers** (fixes `nt_rc1.wad`)